+++ Plugins/ConfigMigration/configmigrationwizard.cpp
@@ -11,6 +11,7 @@
 #include "db/dbsqlite3.h"
 #include "services/notifymanager.h"
 #include "services/dbmanager.h"
+#include "themetuner.h"
 
 ConfigMigrationWizard::ConfigMigrationWizard(QWidget *parent, ConfigMigration* cfgMigration) :
     QWizard(parent),
@@ -40,6 +41,7 @@ void ConfigMigrationWizard::accept()
 void ConfigMigrationWizard::init()
 {
     ui->setupUi(this);
+    THEME_TUNER->darkThemeFix(this);
 
 #ifdef Q_OS_MACX
     resize(width() + 150, height());