+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ### 3.3.3
+- BUGFIX: #4012 Fixed Import/Export wizard windows when using Fusion Dark theme.
 - BUGFIX: #4010 Fixed result column sizing for EXPLAIN queries.
 
 ### 3.3.2